Update all_in_one.py
fix first start bug

+++ all_in_one.py
@@ -3,7 +3,7 @@ import signal
 import subprocess
 import time
 
-from utils import to_struct_time
+from utils import to_struct_time, Config
 
 if __name__ == '__main__':
     fetcher = "python fetch_papers.py\n"
@@ -14,7 +14,9 @@ if __name__ == '__main__':
     thumbnail = "python thumb_pdf.py"
     tmp_script_file = "tmp_script.py"
 
-    download_process = subprocess.Popen(downloader, shell=True)
+    download_process = None
+    if os.path.exists(Config.db_path):
+        download_process = subprocess.Popen(downloader, shell=True)
 
     while True:
         if to_struct_time(time.localtime()).tm_hour == 12:  # exec async tmp file every day at 12:00
@@ -26,8 +28,10 @@ if __name__ == '__main__':
         if to_struct_time(time.localtime()).tm_hour == 14:  # restart download at 14:00
             if os.path.exists(tmp_script_file):
                 os.remove(tmp_script_file)
-            download_process.send_signal(signal.CTRL_C_EVENT)
-            time.sleep(30)  # waiting 30 seconds of terminate time
-            download_process = subprocess.Popen(fetcher, shell=True)
+            if downloader is not None:
+                download_process.send_signal(signal.CTRL_C_EVENT)
+                time.sleep(30)  # waiting 30 seconds of terminate time
+            if os.path.exists(Config.db_path):
+                download_process = subprocess.Popen(fetcher, shell=True)
         subprocess.Popen(thumbnail, creationflags=subprocess.CREATE_NEW_CONSOLE)
         time.sleep(45 * 60)  # sync every 40 minutes
